+++ rpa-pc269/WellTests.cs
@@ -10,15 +10,15 @@ namespace rpa_functions.rpa_pc269
         public int DailyreportId { get; set; }
         public string WellName { get; set; }
         public DateTime? TestDate { get; set; }
-        public short? OnlineTime { get; set; }
+        public decimal? OnlineTime { get; set; }
         public decimal? ChokeOpening { get; set; }
         public decimal? OilProdRate { get; set; }
         public decimal? GasProdRate { get; set; }
         public decimal? WaterProdRate { get; set; }
-        public short? Bhp { get; set; }
-        public short? Bht { get; set; }
-        public short? Whp { get; set; }
-        public short? Wht { get; set; }
+        public decimal? Bhp { get; set; }
+        public decimal? Bht { get; set; }
+        public decimal? Whp { get; set; }
+        public decimal? Wht { get; set; }
         public decimal? SeparatorPressure { get; set; }
         public decimal? Bsw { get; set; }
         public decimal? Gor { get; set; }
